yolo_realtime/requirements.txt
pip==21.3.1
requests==2.27.1
seaborn==0.11.2
PyYAML==6.0
opencv-python==4.1.2.30
torch==1.10.2
torchvision==0.11.3
Pillow==7.1.2
tqdm==4.64.0
matplotlib==3.3.4
pandas==1.1.5